refactor: Add App catalog and put Theme in separate file
