+++ MovieQuiz/App/Theme.swift
@@ -1,0 +1,28 @@
+//
+//  Theme.swift
+//  MovieQuiz
+//
+//  Created by SERGEY SHLYAKHIN on 30.10.2022.
+//
+
+import UIKit
+
+enum Theme {
+    static let boldLargeFont = UIFont(name: "YSDisplay-Bold", size: 23)
+    static let boldSmallFont = UIFont(name: "YSDisplay-Bold", size: 18)
+    static let mediumLargeFont = UIFont(name: "YSDisplay-Medium", size: 20)
+    static let mediumSmallFont = UIFont(name: "YSDisplay-Medium", size: 16)
+    
+    static let buttonCornerRadius: CGFloat = 15
+    static let buttonHeight: CGFloat = 60
+    
+    static let imageCornerRadius: CGFloat = 20
+    static let imageAnswerBorderWidht: CGFloat = 8
+    static let imageHeightAspect: CGFloat = 3/2
+    
+    static let spacing: CGFloat = 20
+    static let leftOffset: CGFloat = 20
+    static let topOffset: CGFloat = 10
+    static let leftQuestionPadding: CGFloat = 42
+    static let topQuestionPadding: CGFloat = 13
+}
