+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -261,24 +261,3 @@ extension MovieQuizViewController {
         showAnswerResult(isCorrect: false == currentQuestion.correctAnswer)
     }
 }
-
-// MARK: - Theme
-enum Theme {
-    static let boldLargeFont = UIFont(name: "YSDisplay-Bold", size: 23)
-    static let boldSmallFont = UIFont(name: "YSDisplay-Bold", size: 18)
-    static let mediumLargeFont = UIFont(name: "YSDisplay-Medium", size: 20)
-    static let mediumSmallFont = UIFont(name: "YSDisplay-Medium", size: 16)
-    
-    static let buttonCornerRadius: CGFloat = 15
-    static let buttonHeight: CGFloat = 60
-    
-    static let imageCornerRadius: CGFloat = 20
-    static let imageAnswerBorderWidht: CGFloat = 8
-    static let imageHeightAspect: CGFloat = 3/2
-    
-    static let spacing: CGFloat = 20
-    static let leftOffset: CGFloat = 20
-    static let topOffset: CGFloat = 10
-    static let leftQuestionPadding: CGFloat = 42
-    static let topQuestionPadding: CGFloat = 13
-}
